requirements.txt
crewai 
langchain-groq
litellm
python-dotenv